+++ src/index.ts
@@ -1,4 +1,10 @@
-export { motion, useExternalRef, createMotionComponent } from "./motion"
+export {
+    motion,
+    useExternalRef,
+    createMotionComponent,
+    HTMLMotionProps,
+    SVGMotionProps,
+} from "./motion"
 export { useMotionValue } from "./value/use-motion-value"
 export { MotionValue, motionValue } from "./value"
 export { unwrapMotionValue } from "./value/utils/unwrap-value"